Fixed null pointer when deleting batch processes [#1955]

+++ comixed-services/src/main/java/org/comixedproject/service/batch/BatchProcessesService.java
@@ -88,8 +88,10 @@ public class BatchProcessesService {
     jobIds.forEach(
         jobId -> {
           final JobExecution jobExecution = this.jobExplorer.getJobExecution(jobId);
-          log.trace("Deleting job execution: id={}", jobExecution.getJobId());
-          this.jobRepository.deleteJobExecution(jobExecution);
+          if (Objects.nonNull(jobExecution)) {
+            log.trace("Deleting job execution: id={}", jobExecution.getJobId());
+            this.jobRepository.deleteJobExecution(jobExecution);
+          }
         });
     return this.getAllBatchProcesses();
   }
